Simulate a panel dataset.

This function simulates a panel dataset with a given number of units and time periods. The user can specify the distribution of the error term, the number of units, the number of time periods, and the parameters of the model. The function returns a panel dataset in the form of a data.table.

Usage

PanelRegSim(
  seed = 1234,
  sample_size = 10000,
  min_year = 2000,
  max_year = 2010,
  noise_sd = 1,
  panel_model = "iid",
  true_beta = 0.5,
  endog_weight = 0.5,
  MA1_persistence = 0.5,
  AR1_persistence = 0.5,
  true_gamma = c(1),
  true_FE_var = 0,
  return_unobservables = FALSE
)

Arguments

seed

(integer) A seed for the random number generator.

sample_size

(integer) The number of units in the panel dataset.

min_year

(integer) The first year in the panel dataset.

max_year

(integer) The last year in the panel dataset.

noise_sd

(numeric) The standard deviation of the measurement error.

panel_model

(character) The panel model to use. Options are "iid", "FE", "AR1", "MA1".

true_beta

(numeric) The true value of the parameter of interest.

endog_weight

(numeric) The weight of the endogenous variable in the panel model. Must be between 0 and 1.

MA1_persistence

(numeric) The persistence parameter of the MA(1) process, i.e., eps_it = shock_it + MA1_persistence * shock_it-1.

AR1_persistence

(numeric) The persistence parameter of the AR(1) process, i.e., eps_it = AR1_persistence * eps_it-1 + shock_it.

true_gamma

(numeric) The true value of the parameter for the exogenous variables. The first entry corresponds to the intercept. By default, true_gamma = c(1).If length(gamma) > 1, then the function adds exogenous variables to the outcome that are correlated with the first endogenous variable.

true_FE_var

(numeric) The variance of the fixed effects. By default, true_FE_var = 0 so that there are no fixed effects. If true_FE_var > 0, then the function adds fixed effects to the outcome that are correlated with the first endogenous variable.

return_unobservables

(logical) If TRUE, the function returns the full panel dataset, including the unobservable variables. If FALSE, the function returns only the observable variables.

Value

(data.table) A panel dataset.